requirements.txt
aiohttp>=3.9.0
aiosqlite>=0.19.0
beautifulsoup4>=4.12.0
click>=8.1.7
duckduckgo-search>=4.1.1
dnspython>=2.5.0
fastapi>=0.109.0
httpx>=0.26.0
networkx>=3.2.1
numpy>=1.26.0
ollama>=0.1.6
Pillow>=10.2.0
playwright>=1.41.0
pydantic>=2.5.0
python-jose>=3.3.0
python-multipart>=0.0.6
python-whois>=0.8.0
pytesseract>=0.3.10
PyYAML>=6.0.1
rich>=13.7.0
SQLAlchemy[asyncio]>=2.0.25
ujson>=5.9.0
uvicorn>=0.27.0